sdm72dm: Use int16_t for samples. Saves ~ 3k RAM.

+++ software/modules/backend/sdm72dm/sdm72dm.h
@@ -55,12 +55,12 @@ private:
     int samples_last_interval = 0;
     int samples_per_interval = -1;
     uint32_t interval_end_ms = 1000 * 60 * HISTORY_MINUTE_INTERVAL;
-    TF_Ringbuffer<int32_t, 3 * 60 * HISTORY_MINUTE_INTERVAL> interval_samples;
+    TF_Ringbuffer<int16_t, 3 * 60 * HISTORY_MINUTE_INTERVAL> interval_samples;
 
     uint32_t next_modbus_read_deadline = 0;
     uint32_t next_power_history_entry = 0;
     UserData user_data;
-    TF_Ringbuffer<int32_t, HISTORY_HOURS * (60 / HISTORY_MINUTE_INTERVAL) + 1> power_history;
+    TF_Ringbuffer<int16_t, HISTORY_HOURS * (60 / HISTORY_MINUTE_INTERVAL) + 1> power_history;
 
     bool energy_meter_reset_requested;
